fix(common): default app load error
