+++ src/lib/auth.ts
@@ -1,3 +1,4 @@
+import { initializeApp } from 'firebase/app';
 import { 
   getAuth, 
   signInWithRedirect, 
@@ -8,8 +9,22 @@ import {
   User
 } from 'firebase/auth';
 
+// Firebase configuration - moved from main.tsx
+const firebaseConfig = {
+  apiKey: "AIzaSyAk9zQ6zDuLFAxhJCXCklNOiBVQQFo1CD8",
+  authDomain: "ut-dnr-ugs-maps-prod.firebaseapp.com",
+  projectId: "ut-dnr-ugs-maps-prod",
+  storageBucket: "ut-dnr-ugs-maps-prod.firebasestorage.app",
+  messagingSenderId: "328621131372",
+  appId: "1:328621131372:web:be0e38a400bb831f79fa49",
+  measurementId: "G-4HPYLDS1SS"
+};
+
+// Initialize Firebase app
+const app = initializeApp(firebaseConfig);
+
 // Get the auth instance
-export const auth = getAuth();
+export const auth = getAuth(app);
 
 // Configure OIDC provider
 const oidcProvider = new OAuthProvider('oidc.utahid');
