+++ src/main.tsx
@@ -5,10 +5,9 @@ import { Toaster } from '@/components/ui/toaster'
 import { ThemeProvider } from '@/context/theme-provider'
 import { MapProvider } from '@/context/map-provider'
 import { SidebarProvider } from '@/context/sidebar-provider'
-import { AuthProvider } from '@/context/auth-provider' // ADD THIS LINE
+import { AuthProvider } from '@/context/auth-provider'
 import { RouterProvider, createRouter } from '@tanstack/react-router'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import { initializeApp } from 'firebase/app'
 import { getAnalytics } from 'firebase/analytics'
 import { logEvent } from 'firebase/analytics';
 
@@ -18,21 +17,15 @@ import { routeTree } from './routeTree.gen'
 // Create a new router instance
 const router = createRouter({ routeTree })
 
-// Firebase configuration
-const firebaseConfig = {
-  apiKey: "AIzaSyAk9zQ6zDuLFAxhJCXCklNOiBVQQFo1CD8",
-  authDomain: "ut-dnr-ugs-maps-prod.firebaseapp.com",
-  projectId: "ut-dnr-ugs-maps-prod",
-  storageBucket: "ut-dnr-ugs-maps-prod.firebasestorage.app",
-  messagingSenderId: "328621131372",
-  appId: "1:328621131372:web:be0e38a400bb831f79fa49",
-  measurementId: "G-4HPYLDS1SS"
-};
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
-const analytics = getAnalytics(app);
-logEvent(analytics, 'app_initialized');
+// Initialize Firebase Analytics (auth is now handled in lib/auth.ts)
+// Note: Only initialize analytics if you need it, otherwise you can remove this
+let analytics;
+try {
+  analytics = getAnalytics();
+  logEvent(analytics, 'app_initialized');
+} catch (error) {
+  console.warn('Analytics not available:', error);
+}
 
 // Register the router instance for type safety
 declare module '@tanstack/react-router' {
